msvc: fix some errors with build MSVC2015

+++ kfft_math.c
@@ -52,8 +52,8 @@ kfft_math_adamar_cpx(kfft_cpx* Fout, kfft_cpx* Fin, uint32_t size) {
 void
 kfft_math_transpose_cpx(const kfft_cpx* Fin, kfft_cpx* Fout, const uint32_t x, const uint32_t y) {
     for (uint64_t n = 0; n < x * y; n++) {
-        uint32_t i = n / y;
-        uint32_t j = n % y;
+        uint64_t i = n / y;
+        uint64_t j = n % y;
         C_CPY(Fout[n], Fin[x * j + i]);
     }
 }
